refactor: move timechart logic to separate file
